add a check to the scripts to make sure only local envs are affected by anything.

+++ bin/teardown.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 set -eou pipefail
 
+# if we aren't in one of the orbstack/docker-desktop contexts, bail out. basically i want to prevent accidentally un-deploying
+# this stuff to staging (or god help us, prod).
+context=$(kubectl config current-context)
+if [[ "$context" != "orbstack" ]] && [[ "$context" != "docker-desktop" ]]; then
+  echo "You are not in the right kube context, current context is: $context. We want 'orbstack' or 'docker-desktop'"
+  exit 1
+fi
+
 helm uninstall -n proxysql proxysql-cluster
 helm uninstall -n proxysql proxysql-cluster-controller
